backend/requirements.txt
xdk
python-dotenv
sqlalchemy[asyncio]>=2.0
psycopg2-binary
alembic
fastapi
uvicorn[standard]
pydantic>=2.0
pydantic-settings
